app/requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
numpy==1.19.4
oauth2client==4.1.3
opencv-python==4.4.0.46
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyserial==3.5
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.2
